third_party/blink/web_tests/dom/mutation-event-tests/fast/dom/MutationObserver/added-out-of-order-expected.txt

CONSOLE WARNING: Listener added for a 'DOMNodeRemoved' mutation event. This event type is deprecated, and will be removed from this browser VERY soon. Usage of this event listener will cause performance issues today, and represents a large risk of imminent site breakage. Consider using MutationObserver instead. See https://chromestatus.com/feature/5083947249172480 for more information.
Test MutationEvents interfering with MutationObservers: adding nodes 'out of order'

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS mutations.length is 3
PASS mutations[0].addedNodes.length is 0
PASS mutations[0].removedNodes.length is 1
PASS mutations[0].removedNodes[0].tagName is 'SPAN'
PASS mutations[1].addedNodes.length is 1
PASS mutations[1].removedNodes.length is 0
PASS mutations[1].addedNodes[0].tagName is 'DIV'
PASS mutations[2].addedNodes.length is 1
PASS mutations[2].removedNodes.length is 0
PASS mutations[2].addedNodes[0].nodeValue is 'hello world'
PASS successfullyParsed is true

TEST COMPLETE
